fix(fullscreen): Reposition window if on wrong monitor
Sending a window to another monitor with a fullscreen-workspace wasn't
working, because, unlike the other layouts, fullscreen doesn't touch all
tiled windows but only the mru-one, leaving the sent window as-is
instead of moving it to the monitor it was sent to. The fullscreen-layout
was changed to check whether any associated windows are located on other
monitors and repositions them.