fix(popup): Reduce flicker when changing workspace
The HazeOver-window uses WinSetAlwaysOnTop to briefly place a window in
the foreground so the dim overlay will be in front of all other windows.
Doing so when there's a popup shown makes it so that the popup is behind
that window for just a moment, making it flicker. To prevent this, show
the popup after updating the focus indicator.