fix: Possibly still treat as active after hide-close sequence
There seem to be cases where – when closing e.g. an active explorer
window – a "hidden" event occurs first, then a "focus" event according
to z-order and eventually a "destroyed" event. Because of the
focus-switch in between the destroyed window is not the active one
anymore and no window is actively set as the new active one with respects
to focusAfterClose.
This commit adds remembering a hidden window if it was active, and in
case said window gets closed within a specific duration after being
hidden, it'll be treated as a regular close event of the active window.